Update how server distinguishes betweeen glTF and blender assets
Rather than keying on data contained within the .glTF file, server takes
responsibility for all objects added by importing the glTF. Specifically,
put them in a special collection and distinguish nodes based on whether
they are in or out of the collection.

It also cleans up the orientation correction so that the point around
which the imported content gets rotated is explicitly the world origin
(and not any blender-inferred origin).